requirements.txt
Flask==2.1.0
Werkzeug==2.0.3
boto3==1.26.0
python-dotenv==0.19.2
flask-cors==3.0.10